+++ tst/player_test.hpp
@@ -8,10 +8,10 @@ void PlayerTest(TestSuite &t) {
     Test test = Test(__func__);
     Player player = Player();
     player.basic();
-    test.assert(static_cast<int>(player.party[0].id), 0, "Player Creature 1 ID");
-    test.assert(static_cast<int>(player.party[1].id), 1, "Player Creature 2 ID");
-    test.assert(static_cast<int>(player.party[2].id), 2, "Player Creature 3 ID");
-    test.assert(static_cast<int>(player.party[0].level), 31, "Player Creature 1 Level");
+    test.assert(player.party[0].id, 0, "Player Creature 1 ID");
+    test.assert(player.party[1].id, 1, "Player Creature 2 ID");
+    test.assert(player.party[2].id, 2, "Player Creature 3 ID");
+    test.assert(player.party[0].level, 31, "Player Creature 1 Level");
 
     t.addTest(test);
 }
